fix filter components for single horizon
